make conn_analyzer have it's own set_evidence helper instead of using flowalert's
